+++ test/AnimatesSymbol.test.ts
@@ -1,24 +1,20 @@
 import Point from "@arcgis/core/geometry/Point";
 import Graphic from "@arcgis/core/Graphic";
-import CIMSymbol from "@arcgis/core/symbols/CIMSymbol.js";
 import PictureMarkerSymbol from "@arcgis/core/symbols/PictureMarkerSymbol.js";
 import SimpleMarkerSymbol from "@arcgis/core/symbols/SimpleMarkerSymbol";
-import * as cimSymbolUtils from "@arcgis/core/symbols/support/cimSymbolUtils";
 import { afterEach, beforeEach, describe, expect, Mock, test, vitest } from "vitest";
 
-import {
-  AnimatedSymbol,
-  updateCIMSymbolPointMarker,
-  updatePictureMarker,
-  updateSimpleMarker,
-} from "../src/AnimatedSymbol";
+import { AnimatedSymbol } from "../src/AnimatedSymbol";
 import {
   AnimationEasingConfig,
   IAnimatedGraphic,
   IPictureMarkerWithOpacity,
   ISimpleMarkerWithOpacity,
 } from "../src/types";
+import { updateSimpleMarker } from "../src/updateMarkers";
 import { getImageAsBase64 } from "../src/utils/encodeimage";
+
+vitest.mock("../src/updateMarkers");
 
 describe("AnimatedSymbol", () => {
   const simpleMarkerSymbol: __esri.SimpleMarkerSymbol = new SimpleMarkerSymbol({
@@ -107,10 +103,10 @@ describe("AnimatedSymbol", () => {
     beforeEach(() => {
       animatedSymbol = new AnimatedSymbol({
         graphic,
-        easingConfig: { type: "spring" },
+        easingConfig: { type: "spring", options: {} },
         id: "test",
       });
-      (animatedSymbol as any).animateSymbolOnStep = mockAnimateSymbol;
+      (animatedSymbol as any).animateSymbolWithSpringEasing = mockAnimateSymbol;
     });
 
     afterEach(() => {
@@ -121,13 +117,14 @@ describe("AnimatedSymbol", () => {
       animatedSymbol.start({});
       expect(mockAnimateSymbol).toHaveBeenCalledWith(
         {},
+        {},
         updateSimpleMarker // default onStep function using update simple marker.
       );
     });
-    test("should be able to be called with animationProps and acustom onStep function", () => {
+    test("should be able to be called with animationProps and a custom onStep function", () => {
       const mockOnStep = vitest.fn();
       animatedSymbol.start({ onStep: mockOnStep });
-      expect(mockAnimateSymbol).toHaveBeenCalledWith({ onStep: mockOnStep }, mockOnStep);
+      expect(mockAnimateSymbol).toHaveBeenCalledWith({ onStep: mockOnStep }, {}, mockOnStep);
     });
   });
 
@@ -135,7 +132,7 @@ describe("AnimatedSymbol", () => {
     test("the scale and rotation is correct at end of spring animation", async () => {
       const animatedGraphic = AnimatedSymbol.createAnimatedGraphic({
         graphic,
-        easingConfig: { type: "spring" },
+        easingConfig: { type: "spring", options: {} },
         id: "animated-graphic",
         isOverlay: true,
       });
@@ -256,143 +253,6 @@ describe("AnimatedSymbol", () => {
       expect((animatedGraphic.symbol as SimpleMarkerSymbol).size).not.toBe(11);
       expect(mockOnStart).toBeCalledTimes(1);
       expect(mockOnFinish).toBeCalledTimes(0);
-    });
-  });
-});
-
-describe("update Symbol properties for", () => {
-  describe("Picture Marker", () => {
-    const pictureMarkerSymbol: __esri.PictureMarkerSymbol = new PictureMarkerSymbol({
-      url: "https://example.com/image.jpg",
-      width: "64",
-      height: "64",
-    });
-
-    test("is scaled correctly", () => {
-      const symb = updatePictureMarker(0.5, pictureMarkerSymbol, { scale: 2 }, pictureMarkerSymbol);
-      expect(symb.height).toBe(64 + 64 * 0.5);
-    });
-    test("is rotated correctly", () => {
-      const symb = updatePictureMarker(
-        0.5,
-        pictureMarkerSymbol,
-        { rotate: -30 },
-        pictureMarkerSymbol
-      );
-      expect(symb.angle).toBe(-30 * 0.5);
-    });
-    test("is made translucent correctly", async () => {
-      vitest.mock("../src/utils/encodeimage");
-      const encodeImageSpy = vitest.fn().mockImplementation(() => {
-        return "testString";
-      });
-      (getImageAsBase64 as Mock<any, any>).mockImplementation(() => encodeImageSpy());
-
-      const symb = updatePictureMarker(
-        0.5,
-        pictureMarkerSymbol,
-        { opacity: 0.5 },
-        pictureMarkerSymbol
-      );
-
-      expect(symb.url).toContain("opacity='0.75'");
-      expect(symb.url).toContain("href='testString'");
-    });
-  });
-
-  describe("Simple Marker", () => {
-    const simpleMarkerSymbol: __esri.SimpleMarkerSymbol = new SimpleMarkerSymbol({
-      size: 10,
-      color: "red",
-      outline: {
-        color: "red",
-        width: 1,
-      },
-    });
-
-    afterEach(() => {
-      vitest.resetAllMocks();
-    });
-    test("is scaled correctly", () => {
-      const symb = updateSimpleMarker(0.5, simpleMarkerSymbol, { scale: 2 }, simpleMarkerSymbol);
-      expect(symb.size).toBe(10 + 10 * 0.5);
-    });
-    test("is rotated correctly", () => {
-      const symb = updateSimpleMarker(0.5, simpleMarkerSymbol, { rotate: -30 }, simpleMarkerSymbol);
-      expect(symb.angle).toBe(-30 * 0.5);
-    });
-    test("is made translucent correctly", () => {
-      const symb = updateSimpleMarker(
-        0.5,
-        simpleMarkerSymbol,
-        { opacity: 0.5 },
-        simpleMarkerSymbol
-      );
-      expect(symb.color.a).toBe(0.75);
-      expect(symb.outline.color.a).toBe(0.75);
-    });
-  });
-
-  describe("CIM Marker", () => {
-    const CIMMarkerSymbol: __esri.CIMSymbol = new CIMSymbol({
-      data: {
-        type: "CIMSymbolReference",
-        symbol: {
-          type: "CIMPointSymbol",
-          symbolLayers: [
-            {
-              type: "CIMVectorMarker",
-              enable: true,
-              size: 32,
-              frame: {
-                xmin: 0,
-                ymin: 0,
-                xmax: 16,
-                ymax: 16,
-              },
-              markerGraphics: [
-                {
-                  type: "CIMMarkerGraphic",
-                  geometry: {
-                    rings: [
-                      [
-                        [8, 16],
-                        [0, 0],
-                        [16, 0],
-                        [8, 16],
-                      ],
-                    ],
-                  },
-                  symbol: {
-                    type: "CIMPolygonSymbol",
-                    symbolLayers: [
-                      {
-                        type: "CIMSolidStroke",
-                        width: 5,
-                        color: [240, 94, 35, 255],
-                      } as any,
-                    ],
-                  },
-                },
-              ],
-            },
-          ],
-        },
-      },
-    });
-
-    test("is scaled correctly", () => {
-      const symb = updateCIMSymbolPointMarker(0.5, CIMMarkerSymbol, { scale: 2 }, CIMMarkerSymbol);
-      expect(cimSymbolUtils.getCIMSymbolSize(symb)).toBe(32 + 32 * 0.5);
-    });
-    test("is rotated correctly", () => {
-      const symb = updateCIMSymbolPointMarker(
-        0.5,
-        CIMMarkerSymbol,
-        { rotate: 30 },
-        CIMMarkerSymbol
-      );
-      expect(cimSymbolUtils.getCIMSymbolRotation(symb)).toBe(360 - 30 * 0.5);
     });
   });
 });
